+++ core/jvm/src/test/scala/com/ccadllc/cedi/dtrace/TestEmitterRecordingTest.scala
@@ -43,28 +43,28 @@ class TestEmitterRecordingTest extends WordSpec with BeforeAndAfterEach with Mat
       testEmitter.cache.containingAll(spanId(spanRoot.spanId.spanId)) should have size (1)
     }
     "support recording a successful new tracing span with new span ID, name, and string note" in {
-      assertChildSpanRecordedWithNote(salesRegionNote)
+      assertChildSpanRecordedOnTimeWithNote(salesRegionNote)
     }
     "support recording a successful new tracing span with new span ID, name, and boolean note" in {
-      assertChildSpanRecordedWithNote(quarterlySalesGoalReachedNote)
+      assertChildSpanRecordedOnTimeWithNote(quarterlySalesGoalReachedNote)
     }
     "support recording a successful new tracing span with new span ID, name, and long note" in {
-      assertChildSpanRecordedWithNote(quarterlySalesUnitsNote)
+      assertChildSpanRecordedOnTimeWithNote(quarterlySalesUnitsNote)
     }
     "support recording a successful new tracing span with new span ID, name, and double note" in {
-      assertChildSpanRecordedWithNote(quarterlySalesTotalNote)
+      assertChildSpanRecordedOnTimeWithNote(quarterlySalesTotalNote)
     }
     "support recording two new successful tracing spans spawned in parallel using `Parallel` typeclass `Applicative` capabilities with notes" in {
-      assertParallelApplicativeChildSpansRecordedWithNotes(salesRegionNote, quarterlySalesGoalReachedNote)
+      assertParallelApplicativeChildSpansRecordedOnTimeWithNotes(salesRegionNote, quarterlySalesGoalReachedNote)
     }
     "support recording two new successful tracing spans spawned in parallel using `NonEmptyParallel` typeclass `Apply` capabilities with notes" in {
-      assertNonEmptyParallelApplyChildSpansRecordedWithNotes(salesRegionNote, quarterlySalesGoalReachedNote)
+      assertNonEmptyParallelApplyChildSpansRecordedOnTimeWithNotes(salesRegionNote, quarterlySalesGoalReachedNote)
     }
     "support recording two new successful tracing spans spawned in parallel using `Parallel` typeclass `Monad` capabilities with notes" in {
-      assertParallelMonadChildSpansRecordedWithNotes(salesRegionNote, quarterlySalesGoalReachedNote)
+      assertParallelMonadChildSpansRecordedOnTimeWithNotes(salesRegionNote, quarterlySalesGoalReachedNote)
     }
     "support recording two new successful tracing spans spawned in parallel using `NonEmptyParallel` typeclass `FlatMap` capabilities with notes" in {
-      assertNonEmptyParallelFlatMapChildSpansRecordedWithNotes(salesRegionNote, quarterlySalesGoalReachedNote)
+      assertNonEmptyParallelFlatMapChildSpansRecordedOnTimeWithNotes(salesRegionNote, quarterlySalesGoalReachedNote)
     }
     "support recording nested spans" in {
       val testEmitter = new TestEmitter[IO]
@@ -109,71 +109,86 @@ class TestEmitterRecordingTest extends WordSpec with BeforeAndAfterEach with Mat
     }
   }
 
-  private def assertChildSpanRecordedWithNote(note: Note): Unit = {
-    val testEmitter = new TestEmitter[IO]
-    val testTimer = TraceSystem.realTimeTimer[IO]
-    val salesManagementSystem = TraceSystem(testSystemData, testEmitter, testTimer)
-    val spanRoot = Span.root(testTimer, quarterlySalesCalculateSpanName).unsafeRunSync
-    IO {
-      Thread.sleep(5L)
-      note
-    }.newAnnotatedSpan(quarterlyPhillySalesCalculateSpanName) { case Right(n) => Vector(n) }.trace(
-      TraceContext(spanRoot, true, salesManagementSystem)).unsafeRunSync
-    val entriesWithNote = testEmitter.cache.containingAll(
-      parentSpanId(spanRoot.spanId.spanId), spanName(quarterlyPhillySalesCalculateSpanName), note.toString)
-    val entriesWithSpanId = testEmitter.cache.containingAll(spanId(spanRoot.spanId.spanId))
-    entriesWithNote should have size (1)
-    entriesWithSpanId should have size (1)
-    ()
+  private def assertChildSpanRecordedOnTimeWithNote(note: Note): Unit = {
+    def assertWithTimer(testTimer: TraceSystem.Timer[IO]): Unit = {
+      val testEmitter = new TestEmitter[IO]
+      val start = testTimer.time.unsafeRunSync
+      val salesManagementSystem = TraceSystem(testSystemData, testEmitter, testTimer)
+      val spanRoot = Span.root(testTimer, quarterlySalesCalculateSpanName).unsafeRunSync
+      IO {
+        Thread.sleep(5L)
+        note
+      }.newAnnotatedSpan(quarterlyPhillySalesCalculateSpanName) { case Right(n) => Vector(n) }.trace(
+        TraceContext(spanRoot, true, salesManagementSystem)).unsafeRunSync
+      val end = testTimer.time.unsafeRunSync
+      val entriesWithNote = testEmitter.cache.containingAll(
+        parentSpanId(spanRoot.spanId.spanId), spanName(quarterlyPhillySalesCalculateSpanName), note.toString)
+      val entriesWithSpanId = testEmitter.cache.containingAll(spanId(spanRoot.spanId.spanId))
+      entriesWithNote should have size (1)
+      entriesWithSpanId should have size (1)
+      entriesWithNote.foreach(_.tc.currentSpan.startTime.value should (be >= start.value and be <= end.value))
+      entriesWithSpanId.foreach(_.tc.currentSpan.startTime.value should (be >= start.value and be <= end.value))
+      ()
+    }
+    assertWithTimer(TraceSystem.realTimeTimer[IO])
+    assertWithTimer(TraceSystem.monotonicTimer[IO])
   }
 
-  private def assertParallelApplicativeChildSpansRecordedWithNotes(note1: Note, note2: Note): Unit = {
+  private def assertParallelApplicativeChildSpansRecordedOnTimeWithNotes(note1: Note, note2: Note): Unit = {
     implicit val co = IO.contextShift(ExecutionContext.global)
-    assertParallelChildSpansRecordedWithNotes(note1, note2) { (io1: TraceIO[Unit], io2: TraceIO[Unit]) =>
+    assertParallelChildSpansRecordedOnTimeWithNotes(note1, note2) { (io1: TraceIO[Unit], io2: TraceIO[Unit]) =>
       (io1 -> io2).parTupled.void
     }
   }
 
-  private def assertNonEmptyParallelApplyChildSpansRecordedWithNotes(note1: Note, note2: Note): Unit = {
+  private def assertNonEmptyParallelApplyChildSpansRecordedOnTimeWithNotes(note1: Note, note2: Note): Unit = {
     implicit val co = IO.contextShift(ExecutionContext.global)
-    assertParallelChildSpansRecordedWithNotes(note1, note2) { (io1: TraceIO[Unit], io2: TraceIO[Unit]) =>
+    assertParallelChildSpansRecordedOnTimeWithNotes(note1, note2) { (io1: TraceIO[Unit], io2: TraceIO[Unit]) =>
       (io1 -> io2).parTupled(NonEmptyParallel[TraceIO, TraceIO.Par]).void
     }
   }
 
-  private def assertParallelMonadChildSpansRecordedWithNotes(note1: Note, note2: Note): Unit = {
+  private def assertParallelMonadChildSpansRecordedOnTimeWithNotes(note1: Note, note2: Note): Unit = {
     implicit val co = IO.contextShift(ExecutionContext.global)
-    assertParallelChildSpansRecordedWithNotes(note1, note2) { (io1: TraceIO[Unit], io2: TraceIO[Unit]) =>
+    assertParallelChildSpansRecordedOnTimeWithNotes(note1, note2) { (io1: TraceIO[Unit], io2: TraceIO[Unit]) =>
       List(io1, io2).parSequence_
     }
   }
 
-  private def assertNonEmptyParallelFlatMapChildSpansRecordedWithNotes(note1: Note, note2: Note): Unit = {
+  private def assertNonEmptyParallelFlatMapChildSpansRecordedOnTimeWithNotes(note1: Note, note2: Note): Unit = {
     implicit val co = IO.contextShift(ExecutionContext.global)
-    assertParallelChildSpansRecordedWithNotes(note1, note2) { (io1: TraceIO[Unit], io2: TraceIO[Unit]) =>
+    assertParallelChildSpansRecordedOnTimeWithNotes(note1, note2) { (io1: TraceIO[Unit], io2: TraceIO[Unit]) =>
       NonEmptyParallel[TraceIO, TraceIO.Par].parProductR(io1)(io2)
     }
   }
 
-  private def assertParallelChildSpansRecordedWithNotes(
+  private def assertParallelChildSpansRecordedOnTimeWithNotes(
     note1: Note, note2: Note)(parAction: (TraceIO[Unit], TraceIO[Unit]) => TraceIO[Unit]): Unit = {
-    val testEmitter = new TestEmitter[IO]
-    val testTimer = TraceSystem.realTimeTimer[IO]
-    val salesManagementSystem = TraceSystem(testSystemData, testEmitter, testTimer)
-    val spanRoot = Span.root(testTimer, quarterlySalesCalculateSpanName).unsafeRunSync
-    parAction(
-      IO(Thread.sleep(3)).void.newSpan(quarterlyPhillySales1CalculateSpanName, note1),
-      IO(
-        Thread.sleep(5)).void.newSpan(quarterlyPhillySales2CalculateSpanName, note2)).trace(TraceContext(spanRoot, true, salesManagementSystem)).unsafeRunSync
-    val entriesWithNote1 = testEmitter.cache.containingAll(
-      parentSpanId(spanRoot.spanId.spanId), spanName(quarterlyPhillySales1CalculateSpanName), note1.toString)
-    val entriesWithNote2 = testEmitter.cache.containingAll(
-      parentSpanId(spanRoot.spanId.spanId), spanName(quarterlyPhillySales2CalculateSpanName), note2.toString)
-    val entriesWithSpanId = testEmitter.cache.containingAll(spanId(spanRoot.spanId.spanId))
-    entriesWithNote1 should have size (1)
-    entriesWithNote2 should have size (1)
-    entriesWithSpanId should have size (1)
-    ()
+    def assertWithTimer(testTimer: TraceSystem.Timer[IO]): Unit = {
+      val testEmitter = new TestEmitter[IO]
+      val start = testTimer.time.unsafeRunSync
+      val salesManagementSystem = TraceSystem(testSystemData, testEmitter, testTimer)
+      val spanRoot = Span.root(testTimer, quarterlySalesCalculateSpanName).unsafeRunSync
+      parAction(
+        IO(Thread.sleep(3)).void.newSpan(quarterlyPhillySales1CalculateSpanName, note1),
+        IO(
+          Thread.sleep(5)).void.newSpan(quarterlyPhillySales2CalculateSpanName, note2)).trace(TraceContext(spanRoot, true, salesManagementSystem)).unsafeRunSync
+      val end = testTimer.time.unsafeRunSync
+      val entriesWithNote1 = testEmitter.cache.containingAll(
+        parentSpanId(spanRoot.spanId.spanId), spanName(quarterlyPhillySales1CalculateSpanName), note1.toString)
+      val entriesWithNote2 = testEmitter.cache.containingAll(
+        parentSpanId(spanRoot.spanId.spanId), spanName(quarterlyPhillySales2CalculateSpanName), note2.toString)
+      val entriesWithSpanId = testEmitter.cache.containingAll(spanId(spanRoot.spanId.spanId))
+      entriesWithNote1 should have size (1)
+      entriesWithNote2 should have size (1)
+      entriesWithSpanId should have size (1)
+      entriesWithNote1.foreach(_.tc.currentSpan.startTime.value should (be >= start.value and be <= end.value))
+      entriesWithNote2.foreach(_.tc.currentSpan.startTime.value should (be >= start.value and be <= end.value))
+      entriesWithSpanId.foreach(_.tc.currentSpan.startTime.value should (be >= start.value and be <= end.value))
+      ()
+    }
+    assertWithTimer(TraceSystem.realTimeTimer[IO])
+    assertWithTimer(TraceSystem.monotonicTimer[IO])
   }
   private def spanName(name: Span.Name) = s"span-name=$name"
   private def spanId(id: Long) = s"span-id=$id"